+++ src/baseInterfaces.ts
@@ -8,95 +8,44 @@ export interface ISwaggerOptions {
   source?: any
   useStaticMethod?: boolean | undefined
   useCustomerRequestInstance?: boolean | undefined
+  include?: Array<string | IInclude>
 }
 
-export interface ISwaggerSource {
-  swagger: string
-  info: string
-  paths: IPaths
-  securityDefinitions: string
-  definitions: IDefinitions
-  externalDocs: string
-}
-
-export interface IPaths {
-  [url: string]: IRequestUrl
-}
-
-export interface IRequestUrl {
-  [method: string]: IRequestMethod
-}
-
-export interface IRequestMethod {
-  tags: string[]
-  summary: string
-  description: string
-  operationId: string
-  consumes: string[]
-  produces: string[]
-  parameters: IParameter[]
-  responses: {
-    [key: string]: {
-      description: string
-      schema: {
-        '$ref': string,
-        'type'?: string,
-        'items'?: IParameterItems,
-      }
-    }
-  }
-}
-
-export type IParameterIn = 'path' | 'formData' | 'query' | 'body'
-
-export interface IParameter {
-  in: IParameterIn
+export interface IPropDef {
   name: string
-  description: string
-  required: string
-  schema: IParameterSchema
-  items: {
-    type: string
-    $ref: string
-  }
   type: string
-  format: string
+  desc: string
 }
 
-export interface IParameterSchema {
-  $ref: string
-  items?: IParameterItems
-  type: string
+export interface IInclude {
+  [key: string]: string[]
+}
+export interface IClassDef {
+  name: string
+  props: IPropDef[]
+  imports: string[]
+}
+
+export interface IDefinitionClass {
+  value: IClassDef
+  name: string
+}
+
+export interface IDefinitionClasses {
+  [key: string]: IDefinitionClass
 }
 
 
-export interface IParameterItems {
-  type?: string
-  $ref: string
-  items?: IParameterItems
+export interface IEnumDef {
+  name: string
+  enumProps: string
 }
 
-export interface IDefinitions {
-  [key: string]: IDefinition
+export interface IDefinitionEnum {
+  name: string
+  value?: IEnumDef | null | undefined
+  content?: string | null | undefined
 }
-
-export interface IDefinition {
-  required: string[]
-  type: 'object' | 'array'
-  properties: IDefinitionProperties
-  enum: any[],
-  items: IDefinitionProperty
-}
-
-export interface IDefinitionProperties {
-  [key: string]: IDefinitionProperty
-}
-export interface IDefinitionProperty {
-  type: string
-  enum: any[]
-  format: string
-  maxLength: number
-  $ref: string
-  items: IDefinitionProperty
-  description: string
+export interface IDefinitionEnums {
+  [key: string]: IDefinitionEnum
 }