+++ src/index.ts
@@ -2,10 +2,12 @@ import * as fs from 'fs'
 import * as path from 'path'
 import prettier from 'prettier';
 import axios from 'axios';
-import { ISwaggerSource, ISwaggerOptions } from './baseInterfaces'
+import { ISwaggerSource } from './saggerInterfaces'
 import { definitionsCodeGen } from './definitionCodegen'
 import { enumTemplate, classTemplate, serviceHeader, customerServiceHeader, serviceTemplate, requestTemplate } from './template';
 import { requestCodegen } from './requestCodegen';
+import { ISwaggerOptions, IInclude } from './baseInterfaces';
+import { findDeepRefs } from './utils';
 
 const defaultOptions: ISwaggerOptions = {
   serviceNameSuffix: 'Service',
@@ -14,30 +16,10 @@ const defaultOptions: ISwaggerOptions = {
   outputDir: './service',
   fileName: 'index.ts',
   useStaticMethod: true,
-  useCustomerRequestInstance: false
+  useCustomerRequestInstance: false,
+  include: []
 }
 
-
-function writeFile(fileDir: string, name: string, data: any) {
-  if (!fs.existsSync(fileDir)) {
-    fs.mkdirSync(fileDir);
-  }
-  const filename = path.join(fileDir, name)
-  console.log('filename', filename);
-  fs.writeFileSync(filename, data)
-}
-
-function format(text: string) {
-  return prettier.format(text, {
-    "printWidth": 120,
-    "tabWidth": 2,
-    "parser": "typescript",
-    "trailingComma": "none",
-    "jsxBracketSameLine": false,
-    "semi": true,
-    "singleQuote": true
-  })
-}
 
 export async function codegen(params: ISwaggerOptions) {
   console.time('finish')
@@ -62,7 +44,10 @@ export async function codegen(params: ISwaggerOptions) {
     ...defaultOptions,
     ...params
   }
-
+  let apiSource = options.useCustomerRequestInstance
+    ? customerServiceHeader
+    : serviceHeader
+  // TODO: next next next time
   // if (options.multipleFileMode) {
   if (false) {
     const { models, enums } = definitionsCodeGen(swaggerSource.definitions)
@@ -82,43 +67,136 @@ export async function codegen(params: ISwaggerOptions) {
       writeFile(fileDir, item.name, format(text))
     })
 
-  } else {
-    let apiSource = options.useCustomerRequestInstance
-      ? customerServiceHeader
-      : serviceHeader
-    const requestClasses = requestCodegen(swaggerSource.paths)
-
-    Object.entries(requestClasses).forEach(([className, requests]) => {
-      let text = ''
-      requests.forEach(req => {
-        const reqName = options.methodNameMode == "operationId"
-          ? req.operationId
-          : req.name
-        text += requestTemplate(reqName, req.requestSchema, options)
-      })
-      text = serviceTemplate(className + options.serviceNameSuffix, text)
-      apiSource += text
-    })
-
+  }
+  else if (options.include && options.include.length > 0) {
+    let reqSource = ''
+    let defSource = ''
+    let requestClasses = Object.entries(requestCodegen(swaggerSource.paths))
     const { models, enums } = definitionsCodeGen(swaggerSource.definitions)
 
-    Object.values(enums).forEach(item => {
-      const text = item.value
-        ? enumTemplate(item.value.name, item.value.enumProps, options.enumNamePrefix)
-        : item.content || ''
+    let allModel = Object.values(models)
+    let allEnum = Object.values(enums)
 
-      apiSource += text
+    options.include.forEach(item => {
+      let allImport: string[] = []
+      let includeClassName = ''
+      let includeRequests: string[] = null
+      if (Object.prototype.toString.call(item) === '[object String]') {
+        includeClassName = <string>item
+      } else {
+        for (let k of Object.keys(item)) {
+          includeClassName = k
+          includeRequests = (<IInclude>item)[k]
+        }
+
+      }
+      for (let [className, requests] of requestClasses) {
+        if (includeClassName !== className) continue
+        let text = ''
+        for (let req of requests) {
+          const reqName = options.methodNameMode == "operationId"
+            ? req.operationId
+            : req.name
+          if (includeRequests) {
+            if (includeRequests.includes(reqName)) {
+              text += requestTemplate(reqName, req.requestSchema, options)
+              // generate ref definition model
+              let imports = findDeepRefs(req.requestSchema.parsedParameters.imports, allModel, allEnum)
+              allImport = allImport.concat(imports)
+            }
+          } else {
+            text += requestTemplate(reqName, req.requestSchema, options)
+            let imports = findDeepRefs(req.requestSchema.parsedParameters.imports, allModel, allEnum)
+            allImport = allImport.concat(imports)
+          }
+        }
+
+        text = serviceTemplate(className + options.serviceNameSuffix, text)
+        reqSource += text
+      }
+
+
+      allModel.forEach(item => {
+        if (allImport.includes(item.name)) {
+          const text = classTemplate(item.value.name, item.value.props, [])
+          defSource += text
+        }
+      })
+
+      allEnum.forEach(item => {
+        if (allImport.includes(item.name)) {
+          const text = item.value
+            ? enumTemplate(item.value.name, item.value.enumProps, options.enumNamePrefix)
+            : item.content || ''
+          defSource += text
+        }
+      })
+
     })
 
-    Object.values(models).forEach(item => {
-      const text = classTemplate(item.value.name, item.value.props, [])
-      apiSource += text
-    })
-
+    apiSource += reqSource + defSource
     writeFile(options.outputDir || '', options.fileName || '', format(apiSource))
+  }
+  else {
+    try {
+
+      Object.entries(requestCodegen(swaggerSource.paths)).forEach(([className, requests]) => {
+        let text = ''
+        requests.forEach(req => {
+
+          const reqName = options.methodNameMode == "operationId"
+            ? req.operationId
+            : req.name
+          text += requestTemplate(reqName, req.requestSchema, options)
+        })
+        text = serviceTemplate(className + options.serviceNameSuffix, text)
+        apiSource += text
+      })
+
+      const { models, enums } = definitionsCodeGen(swaggerSource.definitions)
+
+      Object.values(models).forEach(item => {
+        const text = classTemplate(item.value.name, item.value.props, [])
+        apiSource += text
+      })
+
+      Object.values(enums).forEach(item => {
+        const text = item.value
+          ? enumTemplate(item.value.name, item.value.enumProps, options.enumNamePrefix)
+          : item.content || ''
+        apiSource += text
+      })
+
+
+      writeFile(options.outputDir || '', options.fileName || '', format(apiSource))
+    } catch (error) {
+      console.log('error', error)
+    }
   }
   if (fs.existsSync('./cache_swagger.json')) {
     fs.unlinkSync('./cache_swagger.json');
   }
   console.timeEnd('finish')
+}
+
+
+function writeFile(fileDir: string, name: string, data: any) {
+  if (!fs.existsSync(fileDir)) {
+    fs.mkdirSync(fileDir);
+  }
+  const filename = path.join(fileDir, name)
+  console.log('filename', filename);
+  fs.writeFileSync(filename, data)
+}
+
+function format(text: string) {
+  return prettier.format(text, {
+    "printWidth": 120,
+    "tabWidth": 2,
+    "parser": "typescript",
+    "trailingComma": "none",
+    "jsxBracketSameLine": false,
+    "semi": true,
+    "singleQuote": true
+  })
 }